// src/types.ts
export type KeyIndex = Record<string, number>;
// ...
export interface UserSubstore {
	address: Buffer;
	tokenID: Buffer;
	availableBalance: bigint;
	lockedBalances: {
		module: string;
		amount: bigint;
	}[];
}

export interface UserSubstoreEntry {
	address: string;
	tokenID: string;
	availableBalance: string;
	lockedBalances: {
		module: string;
		amount: string;
	}[];
}

export interface SupplySubstoreEntry {
	tokenID: string;
	totalSupply: string;
}

export interface EscrowSubstoreEntry {
	escrowChainID: Buffer;
	tokenID: Buffer;
	amount: bigint;
}

export interface SupportedTokensSubstoreEntry {
	chainID: Buffer;
	supportedTokenIDs: Buffer[];
}

export interface TokenStoreEntry {
	userSubstore: UserSubstoreEntry[];
	supplySubstore: SupplySubstoreEntry[];
	escrowSubstore: EscrowSubstoreEntry[];
	supportedTokensSubstore: SupportedTokensSubstoreEntry[];
}
// ...
export interface GenesisAssetEntry {
	module: string;
	data: Record<string, unknown>;
	schema: Schema;
}